Fixed chat transcript when agent-client name is the same
The IF statement around line 52 (within the "joined" case) was not recognizing an agent joined chat if the agent name and customer name were the same and therefore the agent would not see chat transcript history and the customer would continue to see messages about waiting for an agent to join (even though the agent had joined).  Replaced the IF statement with a SWITCH statement which evaluates the participantrole instead of evaluating names.